+++ src/main/java/cart/controller/CartApiController.java
@@ -1,50 +1,28 @@
 package cart.controller;
 
-import cart.dto.request.ProductRequestDto;
-import cart.dto.response.ProductResponseDto;
-import cart.service.ProductService;
+import cart.dto.MemberAuthDto;
+import cart.service.CartService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
 import java.net.URI;
-import java.util.List;
 
-@RequestMapping("/products")
 @RestController
-public final class CartApiController {
+@RequestMapping("/carts")
+public class CartApiController {
 
-    private final ProductService productService;
+    private final CartService cartService;
 
-    public CartApiController(final ProductService productService) {
-        this.productService = productService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ProductResponseDto>> findProducts() {
-        final List<ProductResponseDto> result = productService.findProducts();
-        return ResponseEntity.ok(result);
+    public CartApiController(CartService cartService) {
+        this.cartService = cartService;
     }
 
     @PostMapping
-    public ResponseEntity<Void> registerProduct(@Valid @RequestBody ProductRequestDto productRequestDto) {
-        final Long savedProductId = productService.registerProduct(productRequestDto);
-        return ResponseEntity.created(URI.create("/products/" + savedProductId)).build();
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<Void> updateProduct(
-            @PathVariable(name = "id") Long productId,
-            @RequestBody ProductRequestDto productRequestDto
-    ) {
-        productService.updateProduct(productId, productRequestDto);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping(path = "/{id}")
-    public ResponseEntity<Void> removeProduct(@PathVariable(name = "id") Long productId) {
-        productService.removeProduct(productId);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<Void> putInCart(@RequestParam Long productId, MemberAuthDto memberAuthDto) {
+        final Long savedCartId = cartService.putInCart(memberAuthDto, productId);
+        return ResponseEntity.created(URI.create("/carts/" + savedCartId)).build();
     }
 }
-
