+++ src/test/java/cart/controller/CartApiControllerTest.java
@@ -9,142 +9,63 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
+import java.util.Base64;
+
 import static org.hamcrest.Matchers.containsString;
 
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class CartApiControllerTest {
 
     @Autowired
-    ProductService productService;
+    private ProductService productService;
 
     @LocalServerPort
-    int port;
+    private int port;
 
     @BeforeEach
     void setUp() {
         RestAssured.port = port;
     }
 
-    @Test
-    @DisplayName("상품 정보를 받아서 상품을 생성한다.")
-    void create() {
-        //given
-        //when
-        //then
-        RestAssured.given()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(ProductRequestFixture.request)
-                .when()
-                .post("/products")
-                .then()
-                .statusCode(HttpStatus.CREATED.value())
-                .header("Location", containsString("/products/"));
-    }
-
-    @Test
-    @DisplayName("상품의 정보를 수정하여 상품을 업테이트한다.")
-    void update() {
-        //given
-        final Long registeredId = productService.registerProduct(ProductRequestFixture.request);
-
-        //when
-        //then
-        RestAssured.given()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(ProductRequestFixture.updateRequest)
-                .when()
-                .put("/products/" + registeredId)
-                .then()
-                .statusCode(HttpStatus.OK.value());
-    }
-
-    @Test
-    @DisplayName("상품 ID를 통해 상품을 삭제한다.")
-    void delete() {
-        //given
-        final Long registeredId = productService.registerProduct(ProductRequestFixture.request);
-
-        //when
-        //then
-        RestAssured.given()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .delete("/products/" + registeredId)
-                .then()
-                .statusCode(HttpStatus.NO_CONTENT.value());
-    }
-
     @Nested
-    @DisplayName("상품을 생성시 상품 정보가 유효한지 체크한다.")
-    class Create {
+    @DisplayName("장바구니에 상품을 담을 시")
+    class PutInCart {
 
         @Test
-        @DisplayName("이미지 URL이 null 일 때")
-        void nullImageURL() {
-            //given
-            //when
-            //then
+        @DisplayName("멤버, 상품이 유효하다면 장바구니에 추가한다.")
+        void putInCart() {
+            final String encodedAuth = new String(Base64.getEncoder().encode("a@a.com:password1".getBytes()));
+            final Long savedProductId = productService.registerProduct(ProductRequestFixture.request);
+
             RestAssured.given()
                     .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .body(ProductRequestFixture.imageEmptyRequest)
+                    .header("Authorization", "Basic " + encodedAuth)
+                    .queryParam("productId", savedProductId)
                     .when()
-                    .post("/products")
+                    .post("/carts")
                     .then()
-                    .statusCode(HttpStatus.BAD_REQUEST.value())
-                    .body("message", containsString("이미지URL은 비어있을 수 없습니다."));
+                    .statusCode(HttpStatus.CREATED.value())
+                    .header("Location", containsString("/carts/"));
         }
 
         @Test
-        @DisplayName("상품명의 길이가 50자를 초과할 때")
-        void nameOverMaxLength() {
-            //given
-            //when
-            //then
-            RestAssured.given()
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .body(ProductRequestFixture.overLengthNameRequest)
-                    .when()
-                    .post("/products")
-                    .then()
-                    .statusCode(HttpStatus.BAD_REQUEST.value())
-                    .body("message", containsString("상품명은 1이상, 50이하여야 합니다."));
-        }
+        @DisplayName("유효하지 않은 멤버 인증이라면 예외를 던진다.")
+        void putInCartWithInvalidAuth() {
+            final String encodedAuth = new String(Base64.getEncoder().encode("a@a.com".getBytes()));
+            final Long savedProductId = productService.registerProduct(ProductRequestFixture.request);
 
-        @Test
-        @DisplayName("가격이 0보다 작을 때")
-        void priceUnderOne() {
-            //given
-            //when
-            //then
             RestAssured.given()
                     .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .body(ProductRequestFixture.negativePriceRequest)
+                    .header("Authorization", "Basic " + encodedAuth)
+                    .queryParam("productId", savedProductId)
                     .when()
-                    .post("/products")
+                    .post("/carts")
                     .then()
-                    .statusCode(HttpStatus.BAD_REQUEST.value())
-                    .body("message", containsString("가격은 0원 이상 " + Integer.MAX_VALUE + "원 이하여야 합니다."));
-        }
-
-        @Test
-        @DisplayName("중복선택한 카테고리에 없음이 존재할 때")
-        void categoryEmptyUnique() {
-            //given
-            //when
-            //then
-            RestAssured.given()
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .body(ProductRequestFixture.categoryNullRequest)
-                    .when()
-                    .post("/products")
-                    .then()
-                    .statusCode(HttpStatus.BAD_REQUEST.value())
-                    .body("message", containsString("카테고리를 선택해야 합니다."));
+                    .statusCode(HttpStatus.BAD_REQUEST.value());
         }
     }
 }
